Update to move middleware to subpath in src
