+++ index.js
@@ -2,7 +2,7 @@ var async = require('async');
 var mongodb = require('mongodb');
 var DB = require('sharedb').DB;
 var OpLinkValidator = require('./op-link-validator');
-var MiddlewareHandler = require('./src/middlewareHandler');
+var MiddlewareHandler = require('./src/middleware/middlewareHandler');
 
 module.exports = ShareDbMongo;
